Fixes #105
