+++ artisan/job/__init__.py
@@ -1,0 +1,11 @@
+from .base_job import BaseJob
+from .git_job import GitJob
+from .local_job import LocalJob
+from .mercurial_job import MercurialJob
+
+__all__ = [
+    'BaseJob',
+    'GitJob',
+    'LocalJob',
+    'MercurialJob'
+]
